requirements.txt
dash
dash_auth
dash_core_components
dash_html_components
plotly.express
